Retry adding pending RequestManagerFragments.
It's pretty easy to call Glide before super.onCreate in an Activity. If we do so, any attempt we make to add a Fragment will reliably fail. That in turn causes us to register a RequestManager with the Glide singleton that never gets unregistered because the Fragment that should do so is never added.

We're avoiding that here by using the existing pendingFragment maps to verify that the Fragment was in fact added. Previously we just assumed the Fragment was added and removed it from the map in the next frame. Now we check the FragmentManager and loop until either the Fragment is added or the FragmentManager tells us it has been destroyed.

The logic here is duplicated because there's no shared interface across the support and non-support Fragments.

PiperOrigin-RevId: 401324847